Правка №3

+++ controllers/contacts/removeContact.js
@@ -7,21 +7,17 @@ const removeContact = async (req, res, next) => {
   const { id } = req.params;
   const { _id } = req.user;
   // console.log(id);
-  await Contact.findOneAndRemove(
-    { _id: id },
-    { owner: _id },
-    (error, result) => {
-      if (error) {
-        throw new NotFound(`Contact with id=${id} not found`);
-      }
-      res.json({
-        status: "success",
-        code: 200,
-        message: `contact id=${id} delete`,
-        data: { result },
-      });
+  await Contact.findOneAndRemove({ owner: id, _id }, (error, result) => {
+    if (error) {
+      throw new NotFound(`Contact with id=${id} not found`);
     }
-  );
+    res.json({
+      status: "success",
+      code: 200,
+      message: `contact id=${id} delete`,
+      data: { result },
+    });
+  });
 };
 
 // const removeContact = async (req, res, next) => {
